requirements.txt
# YouTube API
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# HTTP requests
requests==2.31.0

# Lemmy API
pythorhead==0.14.1  # Python client for Lemmy API

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pyyaml==6.0.1